backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
feedparser==6.0.11
PyYAML==6.0.2
python-multipart==0.0.9
jinja2==3.1.4
itsdangerous==2.2.0
starlette==0.37.2
